ClientArrayFuture implements Future
